Duplicate check using primary keys.
A little weird, some interactions with the appending of extra information to already existing entries (it only checks primary keys so it thinks it's a duplicate) causing some warnings to show up on the terminal when running testing (test cases with no duplicates).

Some work needs to be done to it, but overall it works.