Updated Javadocs again and tried making a jar file to make a release, getting an error tho.
"Invalid signature file digest for Manifest main attributes"

the commit already has some of the fixes that I found online for this error, but none of them worked.

We can try fixing it later or simply launch with source files only and tell the users to build on their end.